+++ src/main/java/memethespire/MemeTheSpire.java
@@ -1,6 +1,7 @@
 package memethespire;
 
 import basemod.BaseMod;
+import basemod.interfaces.EditKeywordsSubscriber;
 import basemod.interfaces.PostInitializeSubscriber;
 import basemod.interfaces.RelicGetSubscriber;
 import com.evacipated.cardcrawl.modthespire.lib.SpireInitializer;
@@ -8,7 +9,8 @@ import com.megacrit.cardcrawl.relics.AbstractRelic;
 import memethespire.relicmemes.RelicModification;
 
 @SpireInitializer
-public class MemeTheSpire implements PostInitializeSubscriber, RelicGetSubscriber {
+public class MemeTheSpire implements PostInitializeSubscriber,
+        RelicGetSubscriber, EditKeywordsSubscriber {
     public static void initialize() { new MemeTheSpire(); }
 
     public static Config config;
@@ -27,5 +29,12 @@ public class MemeTheSpire implements PostInitializeSubscriber, RelicGetSubscribe
     @Override
     public void receiveRelicGet(AbstractRelic abstractRelic) {
         RelicModification.restore(abstractRelic);
+    }
+
+    @Override
+    public void receiveEditKeywords() {
+        BaseMod.addKeyword(new String[]{"Meme"}, "The description of " +
+                "this card has been modified by a meme! Right-click it to " +
+                "see its original description.");
     }
 }