+++ src/main/java/memethespire/cardmemes/CardModification.java
@@ -64,6 +64,7 @@ public class CardModification {
                             modifiedDescription : modifiedUpgradedDescription) :
                     modifiedDescription;
             card.initializeDescription();
+            card.keywords.add("Meme");
         }
     }